refactor(issue_events): Move <InlineLabel /> into own file
