+++ src/components/index.js
@@ -9,6 +9,7 @@ export * from './issue-description.component';
 export * from './issue-list-item.component';
 export * from './issue-event-list-item.component';
 export * from './label-button.component';
+export * from './inline-label.component';
 export * from './label-list-item.component';
 export * from './github-htmlview.component';
 export * from './markdown-webview.component';
