+++ src/utils/index.js
@@ -2,3 +2,4 @@ export * from './action-helper';
 export * from './loading-animation';
 export * from './text-helper';
 export * from './method-helpers';
+export * from './color-helpers';
